updated tab size on EventModal and refactored file location
